requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
flask==3.1.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
jinja2==3.1.5
markupsafe==3.0.2
packaging==24.2
python-dotenv==1.0.1
pytz==2024.2
sqlalchemy==2.0.36
typing-extensions==4.12.2
werkzeug==3.1.3
